% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{checkConvergence - ELBO Convergence test}
\usage{
checkConvergence(elbo_c, elbo_prev, convergence_threshold)
}
\arguments{
\item{elbo_c}{Current ELBO value.}

\item{elbo_prev}{Previous ELBO value.}

\item{convergence_threshold}{Numeric convergence threshold.}
}
\value{
Logical; TRUE if converged.
}
\description{
Checks whether the algorithm has converged based on ELBO changes.
}
\keyword{internal}
