% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers - psi.R
\name{computeCovMatrix}
\alias{computeCovMatrix}
\title{computeCovMatrix - Gaussian Process Covariance Matrix}
\usage{
computeCovMatrix(Xt, Xp, sigma = 1, w = 1)
}
\arguments{
\item{Xt}{Numeric vector of time points.}

\item{Xp}{Numeric vector of time points.}

\item{sigma}{Numeric variance parameter.}

\item{w}{Numeric decay parameter.}
}
\value{
A covariance matrix of dimension length(Xt) x length(Xp)
}
\description{
Generates a covariance matrix using the specified Gaussian process, scaled by sigma2.
}
\keyword{internal}
