% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers - psi.R
\name{computePsiMatrix}
\alias{computePsiMatrix}
\title{helpers - psi}
\usage{
computePsiMatrix(Xt, Xp, w = 1)
}
\arguments{
\item{Xt}{Numeric vector of time points.}

\item{Xp}{Numeric vector of time points (often identical to Xt).}

\item{w}{Numeric scalar decay parameter, controlling smoothness.}
}
\value{
A covariance matrix of dimension length(Xt) x length(Xp)
}
\description{
Computes Ψ, the correlation matrix, using an exponential kernel.
}
\keyword{internal}
