% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedInvTerms.R
\name{expectedInvTau2}
\alias{expectedInvTau2}
\title{expectedInvTerms - Expected Values for Variance Parameters}
\usage{
expectedInvTau2(lambda1, lambda2)
}
\arguments{
\item{lambda1, lambda2}{Shape and scale parameters for τ².}

\item{delta1, delta2}{Shape and scale parameters for σ².}
}
\value{
Numeric expected values.
}
\description{
Computes the expected inverse and logarithmic terms under the Inverse-gamm distributions for or τ² and σ².
}
\keyword{internal}
