% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedResiduals.R
\name{expectedResidualSq}
\alias{expectedResidualSq}
\title{expectedResiduals - Excpected Squared Residuals and Beta Moments}
\usage{
expectedResidualSq(B, i, y, mu, Sigma, p, iter, psi)
}
\arguments{
\item{B}{List of basis matrices.}

\item{i}{Curve index.}

\item{y}{List of observations.}

\item{mu}{Matrix of posterior beta means.}

\item{Sigma}{Array of posterior beta covariance matrices.}

\item{p}{Matrix of inclusion probabilities.}

\item{iter}{Current iteration index.}

\item{psi}{Correlation matrix.}
}
\value{
Numeric expected squared values.
}
\description{
Computes expected residuals and beta2 moments under the variational posterior
}
\keyword{internal}
